chore: Update TypeScript version to 4.3.4 and compiler to ES2020 (#253)
* chore: Update TypeScript version to 4.3.4

* Refactor some return types from `BigInt` to `bigint`

* Carry `webRtcConnection` forward

+++ ironfish/src/mining/director.test.slow.ts
@@ -358,7 +358,7 @@ describe('Non-fake director tests', () => {
 })
 
 describe('Recalculating target', () => {
-  const minDifficulty = Target.minDifficulty() as bigint
+  const minDifficulty = Target.minDifficulty()
   const strategy = new TestStrategy(new RangeHasher())
   let chain: TestBlockchain
   let memPool: TestMemPool
